backend/fastapi/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pydantic
httpx
llama-cpp-python
mcp
nest-asyncio
python-dotenv
pymongo
arxiv